CF-6 move all components into one component folder
