+++ src/components/BitcoinChart/BitcoinChart.js
