+++ src/components/BitcoinChart/BitcoinChart.styles.js
