+++ src/components/BitcoinChart/index.js
