+++ src/components/Bottom/Bottom.styles.js
