+++ src/components/BottomSkeleton/BottomSkeleton.styles.js
