+++ src/components/ChartsBottom/ChartsBottom.styles.js
