+++ src/components/Currency/Currency.js
