+++ src/components/Currency/Currency.styles.js
