+++ src/components/Currency/index.js
