+++ src/components/Middle/Middle.styles.js
