+++ src/components/ProgressBarChart/ProgressBarChart.js
