+++ src/components/ProgressBarChart/ProgressBarChart.styles.js
