+++ src/components/ProgressBarChart/index.js
