+++ src/components/RightTop/RightTop.styles.js
