+++ src/components/TableCoinSkeleton/TableCoinSkeleton.styles.js
