+++ src/components/VolumeChart/VolumeChart.styles.js
