+++ src/components/index.js
@@ -1,23 +1,23 @@
 export { default as Navbar } from "./Navbar";
-export { default as RightTop } from "./Navbar/RightTop";
-export { default as Top } from "./Navbar/Top";
-export { default as Middle } from "./Navbar/Middle";
-export { default as Bottom } from "./Navbar/Bottom";
-export { default as Currency } from "./Navbar/Currency";
-export { default as SearchBar } from "./Navbar/SearchBar";
-export { default as BottomSkeleton } from "./Navbar/BottomSkeleton";
+export { default as RightTop } from "./RightTop";
+export { default as Top } from "./Top";
+export { default as Middle } from "./Middle";
+export { default as Bottom } from "./Bottom";
+export { default as Currency } from "./Currency";
+export { default as SearchBar } from "./SearchBar";
+export { default as BottomSkeleton } from "./BottomSkeleton";
 export { default as Table } from "./Table";
-export { default as TableSort } from "./Table/TableSort";
-export { default as TableCoin } from "./Table/TableCoin";
-export { default as TableCoinSkeleton } from "./Table/TableCoinSkeleton";
-export { default as ProgressBarChart } from "./Table/ProgressBarChart";
-export { default as TableSparkline } from "./Table/TableSparkline";
+export { default as TableSort } from "./TableSort";
+export { default as TableCoin } from "./TableCoin";
+export { default as TableCoinSkeleton } from "./TableCoinSkeleton";
+export { default as ProgressBarChart } from "./ProgressBarChart";
+export { default as TableSparkline } from "./TableSparkline";
 export { default as Charts } from "./Charts";
-export { default as ChartsBottom } from "./Charts/ChartsBottom";
+export { default as ChartsBottom } from "./ChartsBottom";
 export { default as ValueWithCurrencySymbol } from "./ValueWithCurrencySymbol";
-export { default as BitcoinChart } from "./Charts/BitcoinChart";
-export { default as VolumeChart } from "./Charts/VolumeChart";
-export { default as InfoInfiniteScroll } from "./Table/InfoInfiniteScroll";
+export { default as BitcoinChart } from "./BitcoinChart";
+export { default as VolumeChart } from "./VolumeChart";
+export { default as InfoInfiniteScroll } from "./InfoInfiniteScroll";
 export { default as CoinSummary } from "./CoinSummary";
 export { default as CoinSummarySkeleton } from "./CoinSummarySkeleton";
 export { default as CoinDescription } from "./CoinDescription";
@@ -27,4 +27,4 @@ export { default as FlexSkeleton } from "./FlexSkeleton";
 export { default as CoinConverterLoader } from "./CoinConverterLoader";
 export { default as BackgroundCoinChart } from "./BackgroundCoinChart";
 export { default as BackgroundCoinChartTimeframes } from "./BackgroundCoinChartTimeframes";
-export { default as ProgressBar } from "./Navbar/ProgressBar";
+export { default as ProgressBar } from "./ProgressBar";
